requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.0
click>=8.0.0
scikit-image>=0.19.0
scikit-learn>=1.0.0
scipy>=1.7.0
